Align region and nation detail hooks with shared territory utilities.
Consolidate emissions transforms and header stats, rename conflicting region hooks, and add symmetric page data hooks for region and nation detail pages.
