+++ src/hooks/regions/useRegionKPIs.ts
@@ -10,13 +10,13 @@ type ApiRegionKPI = {
   historicalEmissionChangePercent: number;
 };
 
-export type RegionData = {
+export type RegionKPIData = {
   name: string;
   historicalEmissionChangePercent: number | null;
   meetsParis: boolean | null;
 };
 
-const normalizeRegion = (region: ApiRegionKPI): RegionData => {
+const normalizeRegion = (region: ApiRegionKPI): RegionKPIData => {
   return {
     name: region.region,
     historicalEmissionChangePercent: region.historicalEmissionChangePercent,
@@ -24,7 +24,8 @@ const normalizeRegion = (region: ApiRegionKPI): RegionData => {
   };
 };
 
-export function useRegions() {
+/** Fetches region KPIs from `/regions/kpis` (overview / landing). */
+export function useRegionsKPIs() {
   const {
     data: regionsKPI = [],
     isLoading,
